+++ lib/flint_ui_core.dart
@@ -6,10 +6,12 @@
 library;
 
 export 'src/auth/auth_session.dart';
+export 'src/browser/document.dart';
 export 'src/client/client_router.dart';
 export 'src/component_props.dart' hide toFlintNode;
 export 'src/component.dart';
 export 'src/config/environment_config.dart';
+export 'src/data/resource.dart';
 export 'src/head.dart';
 export 'src/html.dart';
 export 'src/navigation/navigation.dart';
@@ -19,5 +21,6 @@ export 'src/storage/browser_storage.dart';
 export 'src/storage/cookies.dart';
 export 'src/storage/local_storage.dart';
 export 'src/storage/session_storage.dart';
+export 'src/state/state_signal.dart';
 export 'src/style.dart';
 export 'src/widgets.dart';